Move the create dialog in a separate component
