+++ console-frontend/src/pages/apps/mod.rs
@@ -1,7 +1,9 @@
+mod create;
 mod details;
 mod index;
 pub mod ownership;
 
+pub use create::*;
 pub use details::*;
 pub use index::*;
